# specs/openapi.yaml
openapi: 3.1.0
info:
  title: Rabbit-Hole Expression Server API
  version: 1.0.0
  summary: HTTP and WebSocket interface for the Rabbit-Hole EXPRESS layer.
  description: |
    Rabbit-Hole makes AI-agent activity legible. This API is the EXPRESS layer
    over COLLECT telemetry and CLASSIFY semantic flows: it provides health and
    Prometheus observability, session and flow retrieval, structured search,
    natural-language chat, and live flow streaming.

    This contract implements the expression-layer design in
    [S04 — Expression Layer](./04-Expression-Layer.md). Persistent objects and
    their relationships are defined by [S05 — Storage & Data Model](./05-Storage-Data-Model.md).

    Authentication is opt-in. When `RABBITHOLE_API_KEY` is unset, the server
    permits every request. When it is set, routes marked with `ApiKeyAuth`
    require the exact key in the `X-API-Key` header. `/health`, `/metrics`, and
    `/api/v1/metrics` remain public for health probes and Prometheus scrapers.
  contact:
    name: Rabbit-Hole
    url: https://gitlab.readydedis.com/rabbit-hole/rabbit-hole
  license:
    name: MIT
    url: https://opensource.org/license/mit
servers:
  - url: http://127.0.0.1:9734
    description: Default local expression server (RABBITHOLE_LISTEN_ADDR)
    variables: {}
...
paths:
...
  /api/v1/sessions:
    get:
      tags: [Sessions]
      operationId: listSessions
      summary: List monitored sessions
      description: Returns paginated session summaries. The server defaults `offset` to 0 and `limit` to 20, and caps `limit` at 100.
      security:
        - ApiKeyAuth: []
      parameters:
        - $ref: '#/components/parameters/Offset'
        - $ref: '#/components/parameters/SessionListLimit'
      responses:
        '200':
          description: Paginated session summaries.
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestId'
            Access-Control-Allow-Origin:
              $ref: '#/components/headers/CorsAllowOrigin'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/v1/sessions/{id}:
    get:
      tags: [Sessions]
      operationId: getSession
      summary: Get a monitored session
      security:
        - ApiKeyAuth: []
      parameters:
        - $ref: '#/components/parameters/SessionId'
      responses:
        '200':
          description: Full session object, including sanitized process metadata.
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestId'
            Access-Control-Allow-Origin:
              $ref: '#/components/headers/CorsAllowOrigin'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Session'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/SessionNotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/v1/flows/{id}:
    get:
      tags: [Flows]
      operationId: getFlow
      summary: Get a semantic flow
      security:
        - ApiKeyAuth: []
      parameters:
        - $ref: '#/components/parameters/FlowId'
      responses:
        '200':
          description: Full semantic flow, optionally including its captured context window.
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestId'
            Access-Control-Allow-Origin:
              $ref: '#/components/headers/CorsAllowOrigin'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Flow'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/FlowNotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/v1/flows/{id}/context-window:
    get:
      tags: [Flows]
      operationId: getContextWindow
      summary: Get a flow's captured context window
      description: |
        Returns the LLM context captured at the flow's decision point. Context
        capture is optional and disabled by default because it can be expensive.
      security:
        - ApiKeyAuth: []
      parameters:
        - $ref: '#/components/parameters/FlowId'
      responses:
        '200':
          description: Captured context window.
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestId'
            Access-Control-Allow-Origin:
              $ref: '#/components/headers/CorsAllowOrigin'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContextWindow'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/ContextWindowNotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: |
        API key configured through `RABBITHOLE_API_KEY`. Authentication is
        enforced only when that environment variable is non-empty.
  headers:
    RequestId:
      description: Request correlation ID. Echoes the incoming `X-Request-ID` or is generated as a UUID v7.
      schema:
        type: string
        format: uuid
    CorsAllowOrigin:
      description: CORS middleware currently allows all origins.
      schema:
        type: string
        example: '*'
  parameters:
    Offset:
      name: offset
      in: query
      description: Zero-based number of sessions to skip.
      schema:
        type: integer
        format: int32
        minimum: 0
        default: 0
    SessionListLimit:
      name: limit
      in: query
      description: Maximum number of session summaries. Values greater than 100 are capped by the server.
      schema:
        type: integer
        format: int32
        minimum: 1
        maximum: 100
        default: 20
    SessionId:
      name: id
      in: path
      required: true
      description: UUID v7 that identifies a monitored agent session.
      schema:
        $ref: '#/components/schemas/UUIDv7'
    FlowId:
      name: id
      in: path
      required: true
      description: UUID v7 that identifies a semantic flow.
      schema:
        $ref: '#/components/schemas/UUIDv7'
  responses:
...
    Unauthorized:
      description: API key is missing or does not match `RABBITHOLE_API_KEY`.
      headers:
        X-Request-ID:
          $ref: '#/components/headers/RequestId'
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ErrorResponse'}
          example: {error: unauthorized}
    SessionNotFound:
      description: The requested session does not exist.
      headers:
        X-Request-ID:
          $ref: '#/components/headers/RequestId'
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ErrorResponse'}
          example: {error: session not found}
    FlowNotFound:
      description: The requested flow does not exist.
      headers:
        X-Request-ID:
          $ref: '#/components/headers/RequestId'
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ErrorResponse'}
          example: {error: flow not found}
    ContextWindowNotFound:
      description: Context windows are optional, and none is available for this flow.
      headers:
        X-Request-ID:
          $ref: '#/components/headers/RequestId'
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ErrorResponse'}
          example: {error: 'context window not available — may not have been captured'}
...
    InternalServerError:
      description: Unexpected server failure.
      headers:
        X-Request-ID:
          $ref: '#/components/headers/RequestId'
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ErrorResponse'}
          example: {error: internal server error}
  schemas:
    UUIDv7:
      type: string
      format: uuid
      description: UUID version 7 identifier.
      examples: ['0191a000-0000-7000-8000-000000000001']
    RFC3339Timestamp:
      type: string
      format: date-time
      description: UTC timestamp encoded as RFC 3339/RFC 3339 Nano text.
      examples: ['2026-07-21T12:00:00Z']
...
    DurationNanoseconds:
      type: integer
      format: int64
      minimum: 0
      description: |
        Duration encoded as integer nanoseconds — the JSON wire encoding Go's
        `time.Duration` marshals to (json.Marshal of time.Duration emits the
        raw nanosecond count). Used for Flow.duration and
        ContextWindow.duration; GoDuration (string form) remains for fields
        such as HealthResponse.uptime that are emitted via String().
      example: 1200000000
    FlowPhase:
      type: string
      description: Stage in the agent decision cycle (`types.FlowPhase`).
      enum: [observation, deliberation, action, verification]
    FlowOutcome:
      type: string
      description: Result of a semantic flow (`types.FlowOutcome`).
      enum: [success, failure, timeout, unknown]
    SessionStatus:
      type: string
      description: Lifecycle state of a monitored session (`types.SessionStatus`).
      enum: [running, completed, crashed, killed]
...
    SessionMetadata:
      type: object
      additionalProperties: false
      required: [command_line, environment, work_dir, binary_path, version]
      properties:
        command_line:
          type: string
          description: Full command line of the monitored agent process.
        environment:
          type: object
          description: Sanitized environment map; secrets should be redacted before persistence.
          additionalProperties:
            type: string
          example: {OPENAI_API_KEY: '***'}
        work_dir:
          type: string
        binary_path:
          type: string
        version:
          type: string
      example:
        command_line: hermes chat
        environment: {OPENAI_API_KEY: '***'}
        work_dir: /home/kara
        binary_path: /usr/local/bin/hermes
        version: 0.18.2
    Session:
      type: object
      additionalProperties: false
      required: [id, agent_pid, agent_name, start_time, status, metadata]
      properties:
        id:
          $ref: '#/components/schemas/UUIDv7'
        agent_pid:
          type: integer
          format: int32
          description: Operating-system PID of the monitored agent process.
          example: 12345
        agent_name:
          type: string
          examples: [hermes, codex, claude-code]
        start_time:
          $ref: '#/components/schemas/RFC3339Timestamp'
        end_time:
          anyOf:
            - $ref: '#/components/schemas/RFC3339Timestamp'
            - type: 'null'
          description: Null while the session is running.
        status:
          $ref: '#/components/schemas/SessionStatus'
        metadata:
          $ref: '#/components/schemas/SessionMetadata'
      example:
        id: 0191a000-0000-7000-8000-000000000001
        agent_pid: 12345
        agent_name: hermes
        start_time: '2026-07-21T12:00:00Z'
        end_time: null
        status: running
        metadata:
          command_line: hermes chat
          environment: {OPENAI_API_KEY: '***'}
          work_dir: /home/kara
          binary_path: /usr/local/bin/hermes
          version: 0.18.2
    SessionSummary:
      type: object
      additionalProperties: false
      required: [id, agent_pid, agent_name, start_time, status]
      properties:
        id: {$ref: '#/components/schemas/UUIDv7'}
        agent_pid: {type: integer, format: int32}
        agent_name: {type: string}
        start_time: {$ref: '#/components/schemas/RFC3339Timestamp'}
        end_time:
          anyOf:
            - $ref: '#/components/schemas/RFC3339Timestamp'
            - type: 'null'
        status: {$ref: '#/components/schemas/SessionStatus'}
        metadata:
          $ref: '#/components/schemas/SessionMetadata'
    SessionListResponse:
      type: object
      additionalProperties: false
      required: [sessions, total, offset, limit]
      properties:
        sessions:
          type: array
          items: {$ref: '#/components/schemas/SessionSummary'}
        total:
          type: integer
          format: int32
          minimum: 0
          description: Number of summaries returned in this response.
        offset:
          type: integer
          format: int32
          minimum: 0
        limit:
          type: integer
          format: int32
          minimum: 1
          maximum: 100
      example:
        sessions:
          - id: 0191a000-0000-7000-8000-000000000001
            agent_pid: 12345
            agent_name: hermes
            start_time: '2026-07-21T12:00:00Z'
            end_time: null
            status: running
            metadata:
              command_line: hermes chat
              environment: {OPENAI_API_KEY: '***'}
              work_dir: /home/kara
              binary_path: /usr/local/bin/hermes
              version: 0.18.2
        total: 1
        offset: 0
        limit: 20
    ContextWindow:
      type: object
      additionalProperties: false
      required: [flow_id, timestamp, model_name, prompt_text, messages_in, messages_out, token_count, duration]
      properties:
        flow_id:
          $ref: '#/components/schemas/UUIDv7'
        timestamp:
          $ref: '#/components/schemas/RFC3339Timestamp'
        model_name:
          type: string
          example: deepseek-v4-pro
        prompt_text:
          type: string
          description: Complete system prompt visible to the model.
        messages_in:
          type: string
          description: Captured inbound message history.
        messages_out:
          type: string
          description: Captured model response.
        token_count:
          type: integer
          format: int64
          minimum: 0
        duration:
          $ref: '#/components/schemas/DurationNanoseconds'
      example:
        flow_id: 0191b000-0000-7000-8000-000000000001
        timestamp: '2026-07-21T12:00:00Z'
        model_name: deepseek-v4-pro
        prompt_text: 'You are a helpful assistant.'
        messages_in: '[{"role":"user","content":"Read auth.go"}]'
        messages_out: '[{"role":"assistant","content":"..."}]'
        token_count: 2048
        duration: 1500000000
    Flow:
      type: object
      additionalProperties: false
      required:
        - id
        - session_id
        - trace_ids
        - intent
        - phase
        - description
        - outcome
        - confidence
        - start_time
        - end_time
        - duration
        - context_window
        - metadata
      properties:
        id:
          $ref: '#/components/schemas/UUIDv7'
        session_id:
          $ref: '#/components/schemas/UUIDv7'
        trace_ids:
          type: array
          description: UUID v7 identifiers of traces grouped into this flow.
          items: {$ref: '#/components/schemas/UUIDv7'}
        intent:
          type: string
          description: |
            Classifier intent, commonly `read_file`, `patch_code`, `search_web`,
            or `llm_api_call`. Intent is intentionally open-ended.
          examples: [read_file, patch_code, search_web, llm_api_call]
        phase:
          $ref: '#/components/schemas/FlowPhase'
        description:
          type: string
          description: Human-readable description of the classified activity.
        outcome:
          $ref: '#/components/schemas/FlowOutcome'
        confidence:
          type: number
          format: double
          minimum: 0
          maximum: 1
          description: Classifier confidence from 0.0 through 1.0.
        start_time:
          $ref: '#/components/schemas/RFC3339Timestamp'
        end_time:
          $ref: '#/components/schemas/RFC3339Timestamp'
        duration:
          $ref: '#/components/schemas/DurationNanoseconds'
        context_window:
          anyOf:
            - $ref: '#/components/schemas/ContextWindow'
            - type: 'null'
          description: Optional LLM decision-point snapshot.
        metadata:
          type: object
          nullable: true
          description: |
            Arbitrary layer-specific JSON metadata. Null when no metadata was
            recorded for the flow.
          additionalProperties: true
      example:
        id: 0191b000-0000-7000-8000-000000000001
        session_id: 0191a000-0000-7000-8000-000000000001
        trace_ids: [0191c000-0000-7000-8000-000000000001]
        intent: read_file
        phase: action
        description: Read auth.go (247 lines, 2ms)
        outcome: success
        confidence: 0.95
        start_time: '2026-07-21T12:00:00Z'
        end_time: '2026-07-21T12:00:02Z'
        duration: 2000000000
        context_window: null
        metadata: {}
...
    ErrorResponse:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: string
          description: Human-readable error detail.
...